Add support for user not willing to overwrite config file
It's a good practice to let a already existent configuration file untouched,
letting the user to decide if the new tool should or not overwrite. This patch
adds this option via a command line flag specific to the "install" option:

$ ./nipe.pl install -f
...
[.] Overwriting system Tor's config file
[.]   .configs/fedora-torrc -> /etc/tor/torrc

Otherwise, he's notified about the possibility to use one of the configuration
files present in the project's home dir.

$ ./nipe.pl install
...
[.] Refer to our custom Tor config files in project home

NOTE: It can be further enhanced in future version of the tool.